fix: поправил класс Item изменив тип переменной на double feat: запись в конфиг после команды feat: добавил звуки для действий в плагине

+++ src/main/java/ru/fiarr4ik/oceanblockseller/config/Item.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 
         private String name;
         private int amount;
-        private int minPrice;
-        private int maxPrice;
+        private double minPrice;
+        private double maxPrice;
 
     }
